separated the projects informations in a new file and added some design for mobile version of projecs page
